EnvironmentService moved to the Core dir
