+++ app/Services/Core/EnvironmentService.php
@@ -16,7 +16,7 @@
  * Copyright (c) 2019 (original work) MedCenter24.com;
  */
 
-namespace medcenter24\mcCore\App\Services;
+namespace medcenter24\mcCore\App\Services\Core;
 
 
 use medcenter24\mcCore\App\Contract\General\Environment;
